Add enums for `status` fields
This fixes some mistakes I made with the mock arising from the
surprising inconsistencies in Coinbase's API.

The worst situation is the completed statuses - there is "completed" for
orders, "complete" for transactions, and "Complete" for transfers. I
chose to name all three of these enums `complete` to restore sanity.

A few other enum values are named differently from Coinbase's choices
for the same reason.